test: [NUTR-109] - fixes tests with patient mock
